Move comparison lib files into a new subfolder
