+++ src/comparison/cli/main.ts
@@ -1,6 +1,6 @@
-import getFile from "../getFile"
+import getDateFromComparisonFilePath from "../lib/getDateFromComparisonFilePath"
+import getFile from "../lib/getFile"
 import getArgs from "./getArgs"
-import getDateFromComparisonFilePath from "../getDateFromComparisonFilePath"
 import printResult from "./printResult"
 import processFile from "./processFile"
 import processRange from "./processRange"
