+++ src/comparison/lambdas/compareLambda.integration.test.ts
@@ -10,10 +10,10 @@ import MockS3 from "tests/helpers/MockS3"
 import "tests/helpers/setEnvironmentVariables"
 import dynamoDbTableConfig from "tests/helpers/testDynamoDbTableConfig"
 import { ZodError } from "zod"
-import * as CompareMessage from "../compareMessage"
-import createDynamoDbConfig from "../createDynamoDbConfig"
-import createS3Config from "../createS3Config"
 import DynamoGateway from "../DynamoGateway/DynamoGateway"
+import * as CompareMessage from "../lib/compareMessage"
+import createDynamoDbConfig from "../lib/createDynamoDbConfig"
+import createS3Config from "../lib/createS3Config"
 import { isError } from "../Types"
 import type { CompareLambdaEvent } from "../Types/CompareLambdaEvent"
