+++ src/comparison/lambdas/compareLambda.ts
@@ -1,12 +1,12 @@
-import compareMessage from "src/comparison/compareMessage"
-import { isError } from "src/comparison/Types"
-import logger from "src/lib/logging"
-import type { ComparisonResult } from "../compareMessage"
-import createDynamoDbConfig from "../createDynamoDbConfig"
-import createS3Config from "../createS3Config"
+import logger from "../../lib/logging"
 import DynamoGateway from "../DynamoGateway/DynamoGateway"
-import getFileFromS3 from "../getFileFromS3"
-import logInDynamoDb from "../logInDynamoDb"
+import type { ComparisonResult } from "../lib/compareMessage"
+import compareMessage from "../lib/compareMessage"
+import createDynamoDbConfig from "../lib/createDynamoDbConfig"
+import createS3Config from "../lib/createS3Config"
+import getFileFromS3 from "../lib/getFileFromS3"
+import logInDynamoDb from "../lib/logInDynamoDb"
+import { isError } from "../Types"
 import type { CompareLambdaEvent } from "../Types/CompareLambdaEvent"
 import { eventSchema } from "../Types/CompareLambdaEvent"
