+++ src/comparison/lambdas/rerunComparisonLambda.integration.test.ts
@@ -11,10 +11,10 @@ import MockDate from "mockdate"
 import MockDynamo from "tests/helpers/MockDynamo"
 import MockS3 from "tests/helpers/MockS3"
 import dynamoDbTableConfig from "tests/helpers/testDynamoDbTableConfig"
-import createDynamoDbConfig from "../createDynamoDbConfig"
-import createS3Config from "../createS3Config"
 import DynamoGateway from "../DynamoGateway/DynamoGateway"
-import InvokeCompareLambda from "../InvokeCompareLambda"
+import createDynamoDbConfig from "../lib/createDynamoDbConfig"
+import createS3Config from "../lib/createS3Config"
+import InvokeCompareLambda from "../lib/InvokeCompareLambda"
 import type { ComparisonLog } from "../Types"
 import { isError } from "../Types"
 import compareLambda from "./compareLambda"
