+++ src/comparison/lambdas/rerunComparisonLambda.ts
@@ -1,8 +1,8 @@
 import logger from "src/lib/logging"
-import createDynamoDbConfig from "../createDynamoDbConfig"
 import DynamoGateway from "../DynamoGateway/DynamoGateway"
-import getFailedComparisonResults from "../getFailedComparisonResults"
-import InvokeCompareLambda from "../InvokeCompareLambda"
+import createDynamoDbConfig from "../lib/createDynamoDbConfig"
+import getFailedComparisonResults from "../lib/getFailedComparisonResults"
+import InvokeCompareLambda from "../lib/InvokeCompareLambda"
 import { isError } from "../Types"
 
 const { BATCH_SIZE, COMPARISON_LAMBDA_NAME, COMPARISON_BUCKET_NAME } = process.env
