+++ src/comparison/lib/compareMessage.ts
@@ -4,16 +4,16 @@ import isEqual from "lodash.isequal"
 import orderBy from "lodash.orderby"
 import CoreAuditLogger from "src/lib/CoreAuditLogger"
 import type { AnnotatedHearingOutcome } from "src/types/AnnotatedHearingOutcome"
-import generateMockPncQueryResultFromAho from "../../tests/helpers/generateMockPncQueryResultFromAho"
-import getPncQueryTimeFromAho from "../../tests/helpers/getPncQueryTimeFromAho"
-import MockPncGateway from "../../tests/helpers/MockPncGateway"
-import { processTestString } from "../../tests/helpers/processTestFile"
-import { xmlOutputDiff, xmlOutputMatches } from "../comparison/xmlOutputComparison"
-import CoreHandler from "../index"
-import { extractExceptionsFromAho, parseAhoXml } from "../parse/parseAhoXml"
-import convertAhoToXml from "../serialise/ahoXml/generate"
-import type Exception from "../types/Exception"
-import type { Trigger } from "../types/Trigger"
+import generateMockPncQueryResultFromAho from "../../../tests/helpers/generateMockPncQueryResultFromAho"
+import getPncQueryTimeFromAho from "../../../tests/helpers/getPncQueryTimeFromAho"
+import MockPncGateway from "../../../tests/helpers/MockPncGateway"
+import { processTestString } from "../../../tests/helpers/processTestFile"
+import CoreHandler from "../../index"
+import { extractExceptionsFromAho, parseAhoXml } from "../../parse/parseAhoXml"
+import convertAhoToXml from "../../serialise/ahoXml/generate"
+import type Exception from "../../types/Exception"
+import type { Trigger } from "../../types/Trigger"
+import { xmlOutputDiff, xmlOutputMatches } from "./xmlOutputComparison"
 
 type ComparisonResultDebugOutput = {
   triggers: {
