+++ src/comparison/lib/getFailedComparisonResults.ts
@@ -1,5 +1,5 @@
-import type DynamoGateway from "./DynamoGateway/DynamoGateway"
-import type { ComparisonLog, PromiseResult } from "./Types"
+import type DynamoGateway from "../DynamoGateway/DynamoGateway"
+import type { ComparisonLog, PromiseResult } from "../Types"
 
 const getFailedComparisonResults = (dynamoGateway: DynamoGateway, limit: number): PromiseResult<ComparisonLog[]> =>
   dynamoGateway.getFailedOnes(limit)
