+++ src/comparison/lib/logInDynamoDb.integration.test.ts
@@ -1,13 +1,13 @@
-import MockDynamo from "../../tests/helpers/MockDynamo"
-import "tests/helpers/setEnvironmentVariables"
-import createDynamoDbConfig from "./createDynamoDbConfig"
-import DynamoGateway from "./DynamoGateway/DynamoGateway"
-import logInDynamoDb from "./logInDynamoDb"
-import type { ComparisonLog } from "./Types"
-import { isError } from "./Types"
-import dynamoDbTableConfig from "../../tests/helpers/testDynamoDbTableConfig"
 import type { DocumentClient } from "aws-sdk/clients/dynamodb"
 import MockDate from "mockdate"
+import "tests/helpers/setEnvironmentVariables"
+import MockDynamo from "../../../tests/helpers/MockDynamo"
+import dynamoDbTableConfig from "../../../tests/helpers/testDynamoDbTableConfig"
+import DynamoGateway from "../DynamoGateway/DynamoGateway"
+import type { ComparisonLog } from "../Types"
+import { isError } from "../Types"
+import createDynamoDbConfig from "./createDynamoDbConfig"
+import logInDynamoDb from "./logInDynamoDb"
 
 const config = createDynamoDbConfig()
 const dynamoGateway = new DynamoGateway(config)
