+++ src/comparison/lib/runFileOnBichard.ts
@@ -1,8 +1,8 @@
 import fs from "fs"
-import ActiveMqHelper from "../../tests/helpers/ActiveMqHelper"
-import defaults from "../../tests/helpers/defaults"
-import { mockAhoRecordInPnc } from "../../tests/helpers/mockRecordInPnc"
-import type { ImportedComparison } from "./Types/ImportedComparison"
+import ActiveMqHelper from "../../../tests/helpers/ActiveMqHelper"
+import defaults from "../../../tests/helpers/defaults"
+import { mockAhoRecordInPnc } from "../../../tests/helpers/mockRecordInPnc"
+import type { ImportedComparison } from "../Types/ImportedComparison"
 
 const runFileOnBichard = async (comparison: ImportedComparison): Promise<void> => {
   // Insert matching record in PNC
