+++ src/comparison/lib/xmlOutputComparison.ts
